fix when destroying job that is running

+++ app/models/structural.rb
@@ -57,6 +57,7 @@ class Structural < Workflow
   end
 
   def parse_warp3d_log_file
+    return unless staged_dir.exist?
     FileUtils.touch staged_dir.join("update")
     File.open(warp3d_batch_messages_file) do |f|
       lines = f.grep(/new profile/)
